Properly handle extremely large RTTs.

+++ codel5.h
@@ -118,8 +118,8 @@ static inline void kvfree(const void *addr)
 typedef u64 codel_time_t;
 typedef s64 codel_tdiff_t;
 
-#define MS2TIME(a) (a * NSEC_PER_MSEC)
-#define US2TIME(a) (a * NSEC_PER_USEC)
+#define MS2TIME(a) (a * (u64) NSEC_PER_MSEC)
+#define US2TIME(a) (a * (u64) NSEC_PER_USEC)
 
 static inline codel_time_t codel_get_time(void)
 {
